# apps/docs/api-reference/openapi.json
{
  "openapi": "3.1.0",
  "servers": [{ "url": "https://itz.am", "description": "Itzam Live API" }],
  "info": {
    "title": "Itzam API",
    "description": "Development documentation",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Api-Key",
        "description": "The API key to use for authentication"
      }
    },
    "schemas": {
...
      "GetModelsResponse": {
        "type": "object",
        "properties": {
          "models": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "gpt-4o",
                  "description": "The name of the model"
                },
                "tag": {
                  "type": "string",
                  "example": "gpt-4o",
                  "description": "The tag of the model"
                },
                "deprecated": {
                  "type": "boolean",
                  "example": false,
                  "description": "Whether the model is deprecated"
                },
                "hasVision": {
                  "type": "boolean",
                  "example": false,
                  "description": "Whether the model has vision capability"
                },
                "hasReasoningCapability": {
                  "type": "boolean",
                  "example": false,
                  "description": "Whether the model has reasoning capability"
                },
                "isOpenSource": {
                  "type": "boolean",
                  "example": false,
                  "description": "Whether the model is open source"
                },
                "contextWindowSize": {
                  "type": "number",
                  "example": 100000,
                  "description": "The context window size of the model"
                },
                "inputPerMillionTokenCost": {
                  "type": "number",
                  "example": 0.00015,
                  "description": "The input cost per million tokens of the model"
                },
                "outputPerMillionTokenCost": {
                  "type": "number",
                  "example": 0.0006,
                  "description": "The output cost per million tokens of the model"
                },
                "provider": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "OpenAI",
                      "description": "The name of the provider"
                    }
                  },
                  "required": ["name"]
                }
              },
              "required": [
                "name",
                "tag",
                "deprecated",
                "hasVision",
                "hasReasoningCapability",
                "isOpenSource",
                "contextWindowSize",
                "inputPerMillionTokenCost",
                "outputPerMillionTokenCost",
                "provider"
              ]
            }
          }
        },
        "required": ["models"]
      },
      "CreateThreadResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "thread_1234567890",
            "description": "The ID of the created thread"
          },
          "name": {
            "type": "string",
            "example": "My Thread",
            "description": "The name of the thread"
          },
          "lookupKeys": {
            "type": ["array", "null"],
            "items": { "type": "string" },
            "example": ["user-123", "platform-web-app"],
            "description": "The lookup keys of the thread"
          },
          "createdAt": {
            "type": "string",
            "example": "2021-01-01T00:00:00.000Z",
            "description": "The creation date of the thread"
          },
          "updatedAt": {
            "type": "string",
            "example": "2021-01-01T00:00:00.000Z",
            "description": "The last update date of the thread"
          }
        },
        "required": ["id", "name", "lookupKeys", "createdAt", "updatedAt"]
      },
      "CreateThreadInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "My Thread",
            "description": "The name of the thread (optional, will auto-generate if not provided)"
          },
          "lookupKeys": {
            "type": "array",
            "items": { "type": "string" },
            "example": ["user-123", "platform-web-app"],
            "description": "Optional lookup keys for finding the thread later"
          },
          "workflowSlug": {
            "type": "string",
            "minLength": 1,
            "example": "my_great_workflow",
            "description": "The slug of the workflow this thread belongs to"
          }
        },
        "required": ["workflowSlug"]
      },
...
    }
  },
  "security": [{ "ApiKey": [] }],
  "paths": {
...
    "/api/v1/models": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GetModelsResponse" }
              }
            },
            "description": "Successfully retrieved models"
          },
          "400": { "description": "Bad request - invalid input" },
          "401": { "description": "Unauthorized - invalid API key" },
          "403": {
            "description": "Forbidden - organization does not have an active subscription"
          },
          "404": { "description": "Not found - workflow not found" },
          "500": { "description": "Server error during generation" }
        },
        "operationId": "getModels",
        "parameters": [],
        "summary": "Get models",
        "description": "Retrieve all models available on Itzam"
      }
    },
    "/api/v1/threads": {
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateThreadResponse"
                }
              }
            },
            "description": "Successfully created thread"
          },
          "400": { "description": "Bad request - invalid input" },
          "401": { "description": "Unauthorized - invalid API key" },
          "403": {
            "description": "Forbidden - organization does not have an active subscription"
          },
          "404": { "description": "Not found - workflow not found" },
          "500": { "description": "Server error during generation" }
        },
        "operationId": "createThread",
        "parameters": [],
        "summary": "Create thread",
        "description": "Create a new thread for continuous messaging",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateThreadInput" }
            }
          }
        }
      }
    },
...
  }
}